Refactor contact form handling.
